requirements.txt
alembic==1.10.3
altair==4.2.2
asttokens==2.2.1
attrs==23.1.0
azure-ai-formrecognizer==3.3.0b1
azure-common==1.1.28
azure-core==1.26.4
backcall==0.2.0
backoff==2.2.1
black==23.3.0
blinker==1.6.2
blis==0.7.9
boilerpy3==1.0.6
cachetools==5.3.0
canals==0.0.2
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
databricks-cli==0.17.6
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
docker==6.0.1
docopt==0.6.2
elasticsearch==7.17.9
entrypoints==0.4
Events==0.4
executing==1.2.0
farm-haystack==1.15.1
filelock==3.12.0
Flask==2.2.3
fonttools==4.39.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
huggingface-hub==0.14.0
idna==3.4
importlib-metadata==6.6.0
inflect==6.0.4
ipykernel==6.22.0
ipython==8.12.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jarowinkler==1.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
lit==16.0.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==2.3.0
mmh3==3.1.0
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msrest==0.7.1
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
posthog==3.0.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
quantulum3==0.8.1
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
seqeval==1.2.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.10
sqlparse==0.4.4
srsly==2.4.6
sseclient-py==1.7.2
stack-data==0.6.2
streamlit==1.21.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tika==2.6.0
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.25.1
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.2.3
zipp==3.15.0
